requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asn1crypto==1.5.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.10.0
cryptography==42.0.8
fastapi==0.103.0
fastapi-cli==0.0.4
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-secret-manager==2.20.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pg8000==1.31.2
pip==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydantic-settings==2.3.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
scramp==1.4.5
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
sqlmodel==0.0.19
starlette==0.27.0
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.20.0
wheel==0.43.0
yarl==1.9.4
